adding IDs to events
added methods to firebase controllers to get IDs. also clears up the first time sign in code, creating a user for it and setting is unique ID as a field inside of it.  also adjusted the event class to save its ID, and set up a method that can be used to save its id anc create the event. didnt incule parameters because not fully sure on them.