split firebase controller.
to make testing, editing and understanding the code easier, i split the usage of Firebase Controller into 3 parts, user, event and image controller. now anything to do with event(excluding images) are in envents, user(except images) are in user and images in images. adjusted code to edit out fbcontroller and use the new one. kept old one for sake of safety in case something happens. this also allows out firebase stuff to be less bloated, and focus more on one task.