adding firebaseAttendance controller
added firebase attendance controller, handling all attendance stuff. adding more later, but (for now) all the logic to sign someone up or check into an event is there.

note: also added logic to handle if there is a max attendence limit, added that as a int on each event, and added a live count of how many people are currently checked in in a separate subcollection so a listener can listen to just that and not have to load all of the check in subcollection just to see what the count is, as well as it wont trigger the listener every time someone who isnt new checks in.

as of this second anyone who checks in is also signed in, and ill be changing that based on a answer im waiting on in the forums.

next after this is getting signed up displayed.