removed OG controller, encapsulated firebase
removed original firebase controller as it was no longer needed. the removed all calls to firebase directly, instead calling through the controllers. also, changed view events to have a loading ring until all data including the image is loaded. this means we wont see that green square anymore, and also wont see the button flash with different words. as a bonus, if network is slow we wont stare at an unloaded screen but instead a loading sign. i recommend we implement this for all views as it honestly was pretty easy.