limit time, signed up attendees.
i set up time and date limitations to be in firebase so the conference organizers can dynamically change when the events can be held so it can be reused. then added a singleton to old those limits so its only one read a app launch. also set it so that all the times always display in the time zone set by the conference maker (this can be changed? wasnt sure if we want this but did it for now so that if someone from Germany checks the app it will still say 4pm Edmonton time instead of the time it would be in Germany. after that (seriously all that took way too long) i added partial functionality to attendance information button. when clicked, organizer will see a list of all signed up attendees. have not set it up for live updates, and have not finished check in part at all, but what i learned from doing signed up should help that alot. also, added a loading ring before app launch so the singleton for timezones can load because otherwise so many errors. seems to be working now though! i can do the rest sometime this weekend.