events list changes, attending and check in work
changed how browse events works, made it so it only pulls a document that has been changed or added so that it doesnt reload the whole list every time an event is added or changed. should cut down on reads and make the liast load alot faster if theres alot of events. then i finished attendance information. it shows the live count on top of the lists, and has a 2 button navigation displaying the user with thier profile. if you click onchecked in, it shows each user checked in and a sub field of how many times that user has checked in. cna add the geolocation to that later. these two also live update, and show text indicating there are no users signed up.checked in when empty.