added admin priviledges
Admin rights are now restricted. Admin page option is only available if admin field in firebase is true, by default for new users it is false.
Scanning a code assigned to admin auth should make the user an admin upon app restart.
also added a button to see the admin QR code in the admin page